fix float64 in reduceSum

+++ src/Rindow/OpenBLAS/Math_reduction.c
@@ -85,8 +85,8 @@ static PHP_METHOD(Math, reduceSum)
             break;
         }
         case php_interop_polite_math_matrix_dtype_float64: {
-            PHP_RINDOW_OPENBLAS_MATH_DEFDATA_TEMPLATE(float,pDataA,bufferA,offsetA)
-            PHP_RINDOW_OPENBLAS_MATH_DEFDATA_TEMPLATE(float,pDataB,bufferB,offsetB)
+            PHP_RINDOW_OPENBLAS_MATH_DEFDATA_TEMPLATE(double,pDataA,bufferA,offsetA)
+            PHP_RINDOW_OPENBLAS_MATH_DEFDATA_TEMPLATE(double,pDataB,bufferB,offsetB)
             rindow_matlib_d_reducesum((index_t)m,(index_t)n,(index_t)k,pDataA,pDataB);
             break;
         }
